An R Eval

An R Eval is a dataset of challenging R coding problems. Each input is a question about R code which could be solved on first-read only by experts and, with a chance to read documentation and run some code, by fluent data scientists. Solutions are in target() and enable a fluent data scientist to evaluate whether the solution deserves full, partial, or no credit.

Pass this dataset to Task$new() to situate it inside of an evaluation task.

Usage

are

Format

A tibble with 28 rows and 7 columns:

title

Character. Unique identifier/title for the code problem.

input

Character. The question to be answered.

target

Character. The solution, often with a description of notable features of a correct solution.

domain

Character. The technical domain (e.g., Data Analysis, Programming, or Authoring).

task

Character. Type of task (e.g., Debugging, New feature, or Translation.)

source

Character. URL or source of the problem. NAs indicate that the problem was written originally for this eval.

knowledge

List. Required knowledge/concepts for solving the problem.

Source

Posit Community, GitHub issues, R4DS solutions, etc. For row-level references, see source.

Examples

are
#> # A tibble: 28 × 7
#>    title                       input        target domain task  source knowledge
#>    <chr>                       <chr>        <chr>  <chr>  <chr> <chr>  <list>   
#>  1 after-stat-bar-heights      "This bar c… "Pref… Data … New … https… <chr [1]>
#>  2 conditional-grouped-summary "I have a s… "One … Data … New … https… <chr [1]>
#>  3 correlated-delays-reasoning "Here's som… "Nota… Data … New … NA     <chr [1]>
#>  4 curl-http-get               "I have the… "Ther… Progr… Debu… https… <chr [1]>
#>  5 dropped-level-legend        "I'd like t… "Also… Data … New … https… <chr [1]>
#>  6 geocode-req-perform         "I am tryin… "Form… Data … Debu… https… <chr [1]>
#>  7 ggplot-breaks-feature       "Here's a f… "```\… Progr… New … https… <chr [1]>
#>  8 ggplot-breaks-feature       "Here's a f… "```\… Progr… New … https… <chr [1]>
#>  9 grouped-filter-summarize    "Here's som… "Ther… Data … New … NA     <chr [1]>
#> 10 grouped-mutate              "I get an e… "From… Data … Debu… https… <chr [1]>
#> # ℹ 18 more rows

dplyr::glimpse(are)
#> Rows: 28
#> Columns: 7
#> $ title     <chr> "after-stat-bar-heights", "conditional-grouped-summary", "co…
#> $ input     <chr> "This bar chart shows the count of different cuts of diamond…
#> $ target    <chr> "Preferably: \n\n```\nggplot(data = diamonds) + \n  geom_bar…
#> $ domain    <chr> "Data analysis", "Data analysis", "Data analysis", "Programm…
#> $ task      <chr> "New code", "New code", "New code", "Debugging", "New code",…
#> $ source    <chr> "https://jrnold.github.io/r4ds-exercise-solutions/data-visua…
#> $ knowledge <list> "tidyverse", "tidyverse", "tidyverse", "r-lib", "tidyverse"…