Solver: generate output

Solvers evaluate the input in a dataset and produce a result aimed at replicating the target. generate() is the most basic solver in rinspect, and just calls the model with the input and collects the output.

Usage

generate(chat = ellmer::chat_claude())

Arguments

chat

An ellmer chat.

Value

A function that takes an input parameter and an optional chat parameter, returning a named list containing the final result as well as the ellmer chat object that generated it.

Examples

if (interactive()) {
  generate()
}