Compare two task evaluations

Compares performance metrics between two evaluations and calculates statistical measures of their differences.

Usage

inspect_compare(x, y, conf_level = 0.95)

Arguments

x, y

Task objects arising from the same call to Task$new().

conf_level

Optional. A single numeric value between 0 and 1 representing the confidence level. Defaults to 0.95.

Value

A list containing statistics comparing the two task results

mean_x: The mean score of the first task evaluation (x).

mean_y: The mean score of the second task evaluation (y).

mean_diff: The mean difference between scores (x - y).

paired_se: The standard error of the paired differences.

conf_int: A vector containing the lower and upper bounds of the confidence interval for the mean difference.

The function will error if the tasks have different task_ids or different numbers of samples.

Details

The statistics reported here are based on Section 4, Comparing models in "Adding Error Bars to Evals: A Statistical Approach to Language Model Evaluations." Miller (2024), https://arxiv.org/abs/2411.00640.

Notably, confidence intervals are based on Equations (7) and (5) with no clustering variable, and Equation (8) with clustering variables. Epochs are treated as clustering variables, i.e. multiple samples based on a given question are considered clustered.

Examples

if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "") &&
    !identical(Sys.getenv("OPENAI_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  # notably, both of the evaluations passed to `inspect_compare()`
  # will be based on the same call to `Task$new()`
  tsk <- Task$new(
    dataset = simple_addition,
    solver = generate(),
    scorer = model_graded_qa()
  )

  tsk_claude <- tsk$clone()
  tsk_claude$eval(solver_chat = chat_claude())

  tsk_openai <- tsk$clone()
  tsk_openai$eval(solver_chat = chat_openai())

  # do note that this is a rather silly comparison :)
  inspect_compare(tsk_claude, tsk_openai)
}