Write a task to an eval log

This function translates a task to an evaluation log file readable by the Inspect log viewer.

The usual entry point to this function is inspect_view(); use inspect_log() to write a persistent log of your task eval. That said, evaluation logs can't be read back into R as tasks and later analyzed (or viewed in the Inspect log viewer, for that matter); for that use, you likely want to save the task using save() or saveRDS().

Usage

inspect_log(task, dir = attr(res, "dir"))

Arguments

task

An evaluation task created with task_create().

dir

Directory where logs should be stored.

Value

The path to the directory which the log was written to. Pass this value to inspect_view() to view the task logs.