The Inspect Log Viewer

The Inspect Log Viewer

Usage

task_log(task, time_start = Sys.time(), dir = attr(res, "dir"))

inspect_view(x, host = "127.0.0.1", port = 7576, ...)

# S3 method for class 'character'
inspect_view(x, host = "127.0.0.1", port = 7576, ...)

# S3 method for class 'task'
inspect_view(x, host = "127.0.0.1", port = 7576, ...)

Arguments

x

Either a path to a directory containing a task eval log or a task itself. If a task, the function will log the task to a temporary directory and open that directory in the viewer.

host

Host to serve on. Defaults to "127.0.0.1",

port

Port to serve on. Defaults to 7576, one greater than the Python implementation.