Log a function's evaluation

Inspect has a logging mechanism s.t. the names of functions called as well as the arguments supplied are stored in the eval log.

logged() takes a function and maps it to a function with this sort of "logged" return value. The function is applied to both internal functions and user-supplied solvers/scorers/metrics.

Usage

logged(fn, ..., fn_name = deparse(substitute(fn)))

Value

A modified version of fn() with signature ... and a named list return value. That list's return value is:

name: A string representing the name of the function called.

value: The original return value of fn().

arguments: Arguments passed to fn(), captured with rlang::dots_list().

Examples

logged(mean)(x = 1:3)
#> Error in logged(mean): could not find function "logged"