Evaluation tasks

Evaluation tasks provide a flexible data structure for evaluating LLM-based tools.

Datasets contain a set of labelled samples. Datasets are just a tibble with columns input and target, where input is a prompt and target is either literal value(s) or grading guidance. Situate datasets inside of a task with task_create().

Solvers evaluate the input in the dataset and produce a final result. The simplest solver is just an ellmer chat—evaluate a task with a solver using task_solve().

Scorers evaluate the final output of solvers. They may use text comparisons, model grading (like model_graded_qa()), or other custom schemes. Score solver results using task_score().

Usage

task_create(dataset, name = deparse(substitute(dataset)), dir = eval_log_dir())

task_solve(task, solver)

task_score(task, scorer)

Arguments

dataset

A tibble with, minimally, columns input and target.

solver

A function that takes an element of dataset$input as input and determines a value approximating dataset$target. Its return value should be a list with elements result (the final response) and chat (an ellmer chat used to solve the problem, or a list of them). Or, just supply an ellmer chat.

scorer

A function that evaluates how well the solver's return value approximates the corresponding elements of dataset$target.

Value

Each of these functions return a task, which is a subclass of a tibble.

task_create() appends column id.

task_solve() appends columns output and solver.

task_score() appends columns score and scorer.

Examples

if (interactive()) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk
}