Creating tasks

Evaluation tasks provide a flexible data structure for evaluating LLM-based tools.

Datasets contain a set of labelled samples. Datasets are just a tibble with columns input and target, where input is a prompt and target is either literal value(s) or grading guidance. Situate datasets inside of a task with task_create().

Solvers evaluate the input in the dataset and produce a final result. The simplest solver is just an ellmer chat (e.g. ellmer::chat_claude()). Evaluate a task with a solver using task_solve().

Scorers evaluate the final output of solvers. They may use text comparisons (like detect_match()), model grading (like model_graded_qa()), or other custom schemes. Score solver results using task_score().

Usage

task_create(
  dataset,
  name = deparse(substitute(dataset)),
  dir = inspect_log_dir()
)

Arguments

dataset

A tibble with, minimally, columns input and target.

name

A name for the evaluation task. Defaults to deparse(substitute(dataset)).

dir

Directory where logs should be stored.

Value

A task object, which is a subclass of a tibble. task_create() appends column id to its input.

See also

Examples

if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}
#> Scoring ■■■■■■■■■■■■■■■■                  50% | ETA:  2s
#> Scoring ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s