Evaluation tasks

Evaluation tasks provide a flexible data structure for evaluating LLM-based tools.

Usage

task_new(name, dataset, solver, scorer, ...)

task_evaluate(task, ..., dir = eval_log_dir())

Arguments

dataset

A tibble with, minimally, columns input and target.

solver

A function that takes an element of dataset$input as input and determines a value approximating dataset$target. Its return value should be a list with elements result (the final response) and chat (an ellmer chat used to solve the problem, or a list of them).

The rinspect package supplies a number of pre-built solvers; you might start off with generate().

scorer

A function that evaluates how well the solver's return value approximates the corresponding elements of dataset$target.

...

Named task parameters. The resulting task will accept any of these parameters allowing for easily running variants of the task without changing its source code.

Examples