Scoring tasks

Evaluation tasks provide a flexible data structure for evaluating LLM-based tools.

Datasets contain a set of labelled samples. Datasets are just a tibble with columns input and target, where input is a prompt and target is either literal value(s) or grading guidance. Situate datasets inside of a task with task_create().

Solvers evaluate the input in the dataset and produce a final result. The simplest solver is just an ellmer chat—evaluate a task with a solver using task_solve().

Scorers evaluate the final output of solvers. They may use text comparisons (like detect_match()), model grading (like model_graded_qa()), or other custom schemes. Score solver results using task_score().

Usage

task_score(task, scorer)

Arguments

task

An evaluation task created with task_create().

scorer

A function that evaluates how well the solver's return value approximates the corresponding elements of dataset$target. See model-based scoring for examples.

Value

A task object, which is a subclass of a tibble. task_score() appends columns score and scorer to its input.

See also

Examples

if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}
#> Scoring ■■■■■■■■■■■■■■■■                  50% | ETA:  2s
#> Scoring ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s