Interactive evaluation log viewer

rinspect bundles the Inspect log viewer, an interactive app for exploring evaluation logs. You can view tasks directly, which will write the task to a temporary json file, or you can supply a path to a directory of tasks written to json using inspect_log().

Usage

inspect_view(x, host = "127.0.0.1", port = 7576)

# S3 method for class 'character'
inspect_view(x, host = "127.0.0.1", port = 7576)

# S3 method for class 'task'
inspect_view(x, host = "127.0.0.1", port = 7576)

Arguments

x

Either a path to a directory containing a task eval log or a task itself. If a task, the function will log the task to a temporary directory and open that directory in the viewer.

host

Host to serve on. Defaults to "127.0.0.1",

port

Port to serve on. Defaults to 7576, one greater than the Python implementation.

Examples

if (!identical(Sys.getenv("ANTHROPIC_API_KEY"), "")) {
  library(ellmer)
  library(tibble)

  simple_addition <- tibble(
    input = c("What's 2+2?", "What's 2+3?"),
    target = c("4", "5")
  )

  tsk <- task_create(dataset = simple_addition)
  tsk

  tsk <- task_solve(tsk, solver = chat_claude())
  tsk

  tsk <- task_score(tsk, scorer = model_graded_qa())
  tsk

  if (interactive()) {
    inspect_view(tsk)
  }
}
#> Error in req_perform_parallel(reqs, max_active = max_active): HTTP 400 Bad Request.
#> • You have reached your specified API usage limits. You will regain access on 2025-03-01 at 00:00 UTC. [invalid_request_error]