Evaluation tasks

Evaluation tasks provide a flexible data structure for evaluating LLM-based tools.

Usage

task_new(dataset, solver, scorer, ..., name = generate_id())

task_evaluate(task, ...)

Arguments

dataset

A tibble with, minimally, columns input and target.

solver

A function that takes an element of dataset$input as input and determines a value approximating dataset$target. Its return value should be a list with elements result (the final response) and chat (an ellmer chat used to solve the problem, or a list of them).

Alternatively, just supply an ellmer::Chat object!

scorer

A function that evaluates how well the solver's return value approximates the corresponding elements of dataset$target.

...

Named task parameters. The resulting task will accept any of these parameters allowing for easily running variants of the task without changing its source code.

Examples

if (FALSE) { # \dontrun{

library(ellmer)
library(tibble)
library(glue)

dataset <- tibble(
  input = c("What's 2+2?", "What's 2+3?"),
  target = c("4", "5")
)

# requires an ANTHROPIC_API_KEY
solver <- function(input, chat = chat_claude()) {
  ch <- chat$clone()
  res <- ch$chat(input)

  list(result = res, chat = ch)
}

# or, equivalently
solver <- chat_claude()

scorer <- function(input, target, output) {
  ch <- chat_claude()
  response <- ch$chat(glue::glue(
    "An assistant was asked the following: {input}\n",
    "The answer is: {target}.\n",
    "The assistant responded: {output}.\n",
    "Was the assistant correct?\n",
    "End your response with 'Answer: Yes' if yes, 'Answer: No' if no."
  ))

  if (grepl("Answer: Yes", response)) {
    res <- TRUE
  } else if (grepl("Answer: No", response)) {
    res <- FALSE
  } else {
    res <- NA
  }

  list(result = res, chat = ch)
}

task_new(
  dataset = dataset,
  solver = solver,
  scorer = scorer
)
} # }